+++ src/components/__tests__/MetadataServices-test.tsx
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import { stub } from "sinon";
 
 import * as React from "react";
-import { shallow } from "enzyme";
+import { mount } from "enzyme";
 
 import { MetadataServices } from "../MetadataServices";
 
@@ -39,7 +39,7 @@ describe("MetadataServices", () => {
     fetchData = stub();
     editItem = stub().returns(new Promise<void>(resolve => resolve()));
 
-    wrapper = shallow(
+    wrapper = mount(
       <MetadataServices
         data={data}
         fetchData={fetchData}
@@ -54,7 +54,7 @@ describe("MetadataServices", () => {
     let metadataService = wrapper.find("li");
     expect(metadataService.length).to.equal(1);
     expect(metadataService.at(0).text()).to.contain("sample name: test protocol");
-    let editLink = metadataService.at(0).find("a");
+    let editLink = metadataService.at(0).find("a").at(0);
     expect(editLink.props().href).to.equal("/admin/web/config/metadata/edit/2");
   });
 });