+++ src/components/__tests__/TextWithEditMode-test.tsx
@@ -3,6 +3,7 @@ import { stub } from "sinon";
 
 import * as React from "react";
 import { shallow, mount } from "enzyme";
+import { Button } from "library-simplified-reusable-components";
 
 import TextWithEditMode from "../TextWithEditMode";
 import EditableInput from "../EditableInput";
@@ -15,10 +16,9 @@ describe("TextWithEditMode", () => {
   });
 
   it("renders text", () => {
-    let wrapper = shallow(
+    let wrapper = mount(
       <TextWithEditMode text="test" placeholder="editable thing" />
     );
-
     expect(wrapper.text()).to.contain("test");
     expect(wrapper.text()).to.contain("Edit editable thing");
 
@@ -28,34 +28,33 @@ describe("TextWithEditMode", () => {
   });
 
   it("starts in edit mode if there's no text", () => {
-    let wrapper = shallow(
+    let wrapper = mount(
       <TextWithEditMode placeholder="editable thing" />
     );
     let input = wrapper.find(EditableInput);
     expect(input.length).to.equal(1);
     expect(input.prop("placeholder")).to.equal("editable thing");
-    expect(input.prop("value")).to.be.undefined;
-    let link = wrapper.find("a");
-    expect(link.text()).to.contain("Save editable thing");
+    expect(input.prop("value")).to.equal("");
+    let button = wrapper.find(Button);
+    expect(button.text()).to.contain("Save editable thing");
   });
 
   it("switches to edit mode", () => {
-    let wrapper = shallow(
+    let wrapper = mount(
       <TextWithEditMode text="test" placeholder="editable thing" />
     );
     let input = wrapper.find(EditableInput);
     expect(input.length).to.equal(0);
-
-    let editLink = wrapper.find("a");
-    editLink.simulate("click");
+    let editButton = wrapper.find(Button);
+    editButton.simulate("click");
     input = wrapper.find(EditableInput);
     expect(input.length).to.equal(1);
     expect(input.prop("placeholder")).to.equal("editable thing");
     expect(input.prop("value")).to.equal("test");
 
-    let saveLink = wrapper.find("a");
-    expect(saveLink.text()).to.contain("Save");
-    expect(saveLink.text()).not.to.contain("Edit");
+    let saveButton = wrapper.find(Button);
+    expect(saveButton.text()).to.contain("Save");
+    expect(saveButton.text()).not.to.contain("Edit");
   });
 
   it("switches out of edit mode", () => {
@@ -70,8 +69,8 @@ describe("TextWithEditMode", () => {
 
     let getValueStub = stub(EditableInput.prototype, "getValue").returns("new value");
 
-    let saveLink = wrapper.find("a");
-    saveLink.simulate("click");
+    let saveButton = wrapper.find(Button);
+    saveButton.simulate("click");
     input = wrapper.find(EditableInput);
     expect(input.length).to.equal(0);
     expect(onUpdate.callCount).to.equal(1);
@@ -79,9 +78,9 @@ describe("TextWithEditMode", () => {
 
     expect(wrapper.text()).to.contain("new value");
 
-    let editLink = wrapper.find("a");
-    expect(editLink.text()).to.contain("Edit");
-    expect(editLink.text()).not.to.contain("Save");
+    let editButton = wrapper.find(Button);
+    expect(editButton.text()).to.contain("Edit");
+    expect(editButton.text()).not.to.contain("Save");
 
     getValueStub.restore();
   });
@@ -112,8 +111,8 @@ describe("TextWithEditMode", () => {
         onUpdate={onUpdate}
         />
     );
-    let saveLink = wrapper.find("a");
-    saveLink.simulate("click");
+    let saveButton = wrapper.find(Button);
+    saveButton.simulate("click");
     expect(wrapper.text()).to.contain("new value");
     expect(onUpdate.callCount).to.equal(1);
 
@@ -131,10 +130,10 @@ describe("TextWithEditMode", () => {
         onUpdate={onUpdate}
         />
     );
-    let editLink = wrapper.find("a");
-    editLink.simulate("click");
-    saveLink = wrapper.find("a");
-    saveLink.simulate("click");
+    let editButton = wrapper.find(Button);
+    editButton.simulate("click");
+    saveButton = wrapper.find(Button);
+    saveButton.simulate("click");
     expect(wrapper.text()).not.to.contain("test");
     expect(wrapper.text()).to.contain("new value");
     expect(onUpdate.callCount).to.equal(3);