+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 #### Updated
 - Updated `typings` to `@types`, updated Typescript, updated Webpack, and updated unit tests with fetch-mock. These updates are needed to import the updated `opds-web-client` package.
 
+### v0.1.12
+#### Updated
+- Updated `reusable-components` to v1.3.1, bringing in the reusable Button component.
+- Implemented the new Panel customization option from v1.3.2 of `reusable-components`.
+
 ### v0.1.11
 #### Fixed
 - `package-lock.json` was not insynced with the latest build and caused installing and running test issues.